Move code around before separating PR
