+++ src/website/TestCases/Grid/table-8.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom";
 import { MarathonEnvironment } from "../../Marathon";
 
-import DataFrame from "../../../data_handling/DataFrame";
+import DataFrame from "../../../DataFrame/DataFrame";
 import { NewGrid } from "../../../NewGrid/NewGrid";
 
 import AutoSizer from "react-virtualized-auto-sizer";
@@ -76,7 +76,7 @@ const rawData = {
 
 const frame = new DataFrame(rawData.columns, rawData.rows);
 
-import { getQuantitiveStats } from "./../../../data_handling/stats";
+import { getQuantitiveStats } from "../../../DataFrame/stats";
 
 const stats = getQuantitiveStats(frame);
 const max: number = Math.max(stats.max.revenue, stats.max.sales);
